refactors the program form to inject submission logic separately
fixes gql mutation query error check for toasting
